Enable >= 2^31 tokens in input data

+++ basic/mem-tracker.h
@@ -39,14 +39,14 @@
 enum MemType { T_RAWNUMBER, list_types(prefix_t) };
 
 struct MemRecord {
-  MemRecord(const char *name, int mem) :
+  MemRecord(const char *name, long mem) :
     name(name), type(T_RAWNUMBER), data(NULL), mem(mem) { }
   MemRecord(const char *name, MemType type, const void *data) :
     name(name), type(type), data(data), mem(0) { }
   string name;
   MemType type;
   const void *data;
-  int mem;
+  long mem;
 };
 
 // Track amount of memory used.
@@ -56,12 +56,12 @@ public:
 
   list_types(define_add)
 
-  void add(const char *name, int mem) {
+  void add(const char *name, long mem) {
     records.push_back(MemRecord(name, mem));
   }
 
-  int compute_mem_usage(const MemRecord &r);
-  int compute_mem_usage();
+  long compute_mem_usage(const MemRecord &r);
+  long compute_mem_usage();
   void report_mem_usage();
 
 private:
@@ -73,8 +73,8 @@ extern MemTracker mem_tracker;
 ////////////////////////////////////////////////////////////
 // Various mem_usage() functions on various data types.
 
-template<class T> int mem_usage(const vector< vector< vector< vector<T> > > > &mat) { // matrix
-  int mem = 0;
+template<class T> long mem_usage(const vector< vector< vector< vector<T> > > > &mat) { // matrix
+  long mem = 0;
   foridx(i, len(mat)) {
     foridx(j, len(mat[i])) {
       foridx(k, len(mat[i][j]))
@@ -87,8 +87,8 @@ template<class T> int mem_usage(const vector< vector< vector< vector<T> > > > &m
   return mem;
 }
 
-template<class T> int mem_usage(const vector< vector< vector<T> > > &mat) { // matrix
-  int mem = 0;
+template<class T> long mem_usage(const vector< vector< vector<T> > > &mat) { // matrix
+  long mem = 0;
   foridx(i, len(mat)) {
     foridx(j, len(mat[i]))
       mem += len(mat[i][j]) * sizeof(T);
@@ -98,32 +98,32 @@ template<class T> int mem_usage(const vector< vector< vector<T> > > &mat) { // m
   return mem;
 }
 
-template<class T> int mem_usage(const vector< vector<T> > &mat) { // matrix
-  int mem = 0;
+template<class T> long mem_usage(const vector< vector<T> > &mat) { // matrix
+  long mem = 0;
   foridx(i, len(mat))
     mem += len(mat[i]) * sizeof(T);
   mem += len(mat) * sizeof(vector<T>);
   return mem;
 }
 
-template<class T> int mem_usage(const vector<T> &vec) { // vector
+template<class T> long mem_usage(const vector<T> &vec) { // vector
   return len(vec) * sizeof(T);
 }
 
-template<class T> int mem_usage(const unordered_set<T> &set) { // hash_set
-  return (int)set.bucket_count()*4 + len(set)*(sizeof(T)+sizeof(void *));
+template<class T> long mem_usage(const unordered_set<T> &set) { // hash_set
+  return (long)set.bucket_count()*4 + len(set)*(sizeof(T)+sizeof(void *));
 }
 
-template<class Tx, class Ty, class Hf, class Eq> int mem_usage(const unordered_map<Tx, Ty, Hf, Eq> &map) { // hash_map
-  return (int)map.bucket_count()*4 + len(map)*(sizeof(Tx)+sizeof(Ty)+sizeof(void *));
+template<class Tx, class Ty, class Hf, class Eq> long mem_usage(const unordered_map<Tx, Ty, Hf, Eq> &map) { // hash_map
+  return (long)map.bucket_count()*4 + len(map)*(sizeof(Tx)+sizeof(Ty)+sizeof(void *));
 }
 
-inline int mem_usage(const UnionSet &u) { // UnionSet
+inline long mem_usage(const UnionSet &u) { // UnionSet
   return mem_usage(u.parent);
 }
 
-inline int mem_usage(const StrDB &db) { // StrDB
-  int mem = mem_usage(db.s2i) + mem_usage(db.i2s);
+inline long mem_usage(const StrDB &db) { // StrDB
+  long mem = mem_usage(db.s2i) + mem_usage(db.i2s);
   foridx(i, len(db))
     mem += (strlen(db[i])+1) * sizeof(char);
   return mem;
